feat(react-wallet): Move NavDrawer to separate component
